+++ grails-app/services/com/tgid/tri/race/SegmentService.groovy
@@ -2,7 +2,7 @@ package com.tgid.tri.race
 
 class SegmentService {
 
-    Segment findOrSaveSegment(Float distance, DistanceType distanceType, SegmentType segmentType) {
+    Segment findOrSaveSegment(Double distance, DistanceType distanceType, SegmentType segmentType) {
         try {
             Segment.findOrSaveWhere(distance: distance, distanceType: distanceType, segmentType: segmentType)
         } catch(Exception e) {
@@ -11,12 +11,12 @@ class SegmentService {
     }
 
     void seedSegments() {
-        findOrSaveSegment(0.3f, DistanceType.Miles, SegmentType.Swim)
-        findOrSaveSegment(0.5f, DistanceType.Miles, SegmentType.Swim)
-        findOrSaveSegment(0.9f, DistanceType.Miles, SegmentType.Swim)
+        findOrSaveSegment(0.3, DistanceType.Miles, SegmentType.Swim)
+        findOrSaveSegment(0.5, DistanceType.Miles, SegmentType.Swim)
+        findOrSaveSegment(0.9, DistanceType.Miles, SegmentType.Swim)
         findOrSaveSegment(1, DistanceType.Miles, SegmentType.Swim)
-        findOrSaveSegment(1.2f, DistanceType.Miles, SegmentType.Swim)
-        findOrSaveSegment(2.4f, DistanceType.Miles, SegmentType.Swim)
+        findOrSaveSegment(1.2, DistanceType.Miles, SegmentType.Swim)
+        findOrSaveSegment(2.4, DistanceType.Miles, SegmentType.Swim)
         findOrSaveSegment(2, DistanceType.Miles, SegmentType.Swim)
         findOrSaveSegment(3, DistanceType.Miles, SegmentType.Swim)
         findOrSaveSegment(4, DistanceType.Miles, SegmentType.Swim)
@@ -24,7 +24,7 @@ class SegmentService {
 
         findOrSaveSegment(11, DistanceType.Miles, SegmentType.Bike)
         findOrSaveSegment(13, DistanceType.Miles, SegmentType.Bike)
-        findOrSaveSegment(13.5f, DistanceType.Miles, SegmentType.Bike)
+        findOrSaveSegment(13.5, DistanceType.Miles, SegmentType.Bike)
         findOrSaveSegment(15, DistanceType.Miles, SegmentType.Bike)
         findOrSaveSegment(20, DistanceType.Kilometers, SegmentType.Bike)
         findOrSaveSegment(25, DistanceType.Kilometers, SegmentType.Bike)
@@ -37,14 +37,14 @@ class SegmentService {
 
         findOrSaveSegment(1, DistanceType.Miles, SegmentType.Run)
         findOrSaveSegment(3, DistanceType.Miles, SegmentType.Run)
-        findOrSaveSegment(3.1f, DistanceType.Miles, SegmentType.Run)
+        findOrSaveSegment(3.1, DistanceType.Miles, SegmentType.Run)
         findOrSaveSegment(5, DistanceType.Kilometers, SegmentType.Run)
         findOrSaveSegment(4, DistanceType.Miles, SegmentType.Run)
         findOrSaveSegment(5, DistanceType.Miles, SegmentType.Run)
         findOrSaveSegment(8, DistanceType.Kilometers, SegmentType.Run)
         findOrSaveSegment(10, DistanceType.Kilometers, SegmentType.Run)
         findOrSaveSegment(25, DistanceType.Kilometers, SegmentType.Run)
-        findOrSaveSegment(13.1f, DistanceType.Miles, SegmentType.Run)
-        findOrSaveSegment(26.2f, DistanceType.Miles, SegmentType.Run)
+        findOrSaveSegment(13.1, DistanceType.Miles, SegmentType.Run)
+        findOrSaveSegment(26.2, DistanceType.Miles, SegmentType.Run)
     }
 }
